Edicion de eliminar promocion

+++ src/controllers/productos.controllers.js
@@ -136,15 +136,20 @@ productosCtrl.actualizarPromocion = async (req, res) => {
 productosCtrl.eliminarPromocion = async (req, res) => {
 	try {
 		const promocionBase = await promocionModel.findById(req.params.id);
-		if (promocionBase.imagenPromocion) {
-			await imagen.eliminarImagen(promocionBase.imagenPromocion);
-		}
-	
-		const promocion = await promocionModel.findByIdAndDelete(req.params.id);
-		if (!promocion) {
+		if(promocionBase){
+			if (promocionBase.imagenPromocion) {
+				await imagen.eliminarImagen(promocionBase.imagenPromocion);
+			}
+		
+			const promocion = await promocionModel.findByIdAndDelete(req.params.id);
+			if (!promocion) {
+				res.status(404).json({ message: 'Este promocion no existe' });
+			}
+			res.status(200).json({ message: 'Promocion eliminada' });
+		}else{
 			res.status(404).json({ message: 'Este promocion no existe' });
 		}
-		res.status(200).json({ message: 'Promocion eliminada' });
+
 	} catch (err) {
 		res.status(500).json({ message: "Error en el servidor",err })	
 	}
